+++ requirements.py
@@ -1,4 +1,6 @@
 nose
 mock
-oauth2
-urllib3
+ipython
+oauth2==1.5.211
+urllib3==1.10
+httplib2.system-ca-certs-locater==0.1.2